+++ R/comorbidity.R
@@ -8,13 +8,24 @@
 #' @param id Column of `x` containing the individual ID.
 #' @param code Column of `x` containing diagnostic codes. Codes must be in upper case with no punctuation in order to be properly recognised.
 #' @param score The comorbidity score to compute. Possible choices are the weighted Charlson score (`charlson`) and the weighted Elixhauser score (`elixhauser`). Values are case-insensitive.
-#' @param assign0 Apply a hierarchy of comorbidities. If `TRUE`, should a comorbidity be present in a patient with different degrees of severity, then the milder form will be assigned to 0 and therefore not counted. By doing this, a type of comorbidity is not counted more than once in each patient. In particular, the comorbidities that are affected by this argument are:
+#' @param assign0 Apply a hierarchy of comorbidities: should a comorbidity be present in a patient with different degrees of severity, then the milder form will be assigned a value of 0.
+#' By doing this, a type of comorbidity is not counted more than once in each patient.
+#'
+#' If `assign0 = "score"` then the hierarchy will be applied to the score and weighted score only, i.e. the comorbidity domains will not be affected.
+#' If `assign0 = "both"` then both domains and scores will be affected.
+#' Finally, if `assign0 = "none"` no hierarchy is applied.
+#' The previous behaviour (`comorbidity <= 0.5.3`) corresponds to:
+#' * `assign0 = "score"` replaces `assign0 = TRUE`,
+#' * `assign0 = "none"` replaces `assign0 = FALSE`.
+#'
+#' The comorbidities that are affected by this argument are:
 #' * "Mild liver disease" (`mld`) and "Moderate/severe liver disease" (`msld`) for the Charlson score;
 #' * "Diabetes" (`diab`) and "Diabetes with complications" (`diabwc`) for the Charlson score;
 #' * "Cancer" (`canc`) and "Metastatic solid tumour" (`metacanc`) for the Charlson score;
 #' * "Hypertension, uncomplicated" (`hypunc`) and "Hypertension, complicated" (`hypc`) for the Elixhauser score;
 #' * "Diabetes, uncomplicated" (`diabunc`) and "Diabetes, complicated" (`diabc`) for the Elixhauser score;
 #' * "Solid tumour" (`solidtum`) and "Metastatic cancer" (`metacanc`) for the Elixhauser score.
+#'
 #' @param icd The version of ICD coding to use. Possible choices are ICD-9-CM (`icd9`) or ICD-10 (`icd10`). Defaults to `icd10`, and values are case-insensitive.
 #' @param factorise Return comorbidities as factors rather than numeric, where (1 = presence of comorbidity, 0 = otherwise). Defaults to `FALSE`.
 #' @param labelled Attach labels to each comorbidity, compatible with the RStudio viewer via the [utils::View()] function. Defaults to `TRUE`.
@@ -106,13 +117,24 @@
 #' )
 #'
 #' # Charlson score based on ICD-10 diagnostic codes:
-#' comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
+#' comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
 #'
 #' # Elixhauser score based on ICD-10 diagnostic codes:
-#' comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
+#' comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = "none")
 #' @export
 
 comorbidity <- function(x, id, code, score, assign0, icd = "icd10", factorise = FALSE, labelled = TRUE, tidy.codes = TRUE) {
+
+  # x = x
+  # id = "id"
+  # code = "code"
+  # score = "elixhauser"
+  # assign0 = "both"
+  # icd = "icd10"
+  # factorise = FALSE
+  # labelled = TRUE
+  # tidy.codes = TRUE
+
   ### Check arguments
   arg_checks <- checkmate::makeAssertCollection()
   # x must be a data.frame (or a data.table)
@@ -122,14 +144,16 @@ comorbidity <- function(x, id, code, score, assign0, icd = "icd10", factorise = 
   checkmate::assert_string(code, add = arg_checks)
   checkmate::assert_string(score, add = arg_checks)
   checkmate::assert_string(icd, add = arg_checks)
+  # assign0 must be a string with possible values: 'score', 'both', 'none'
+  checkmate::assert_string(assign0, add = arg_checks)
+  checkmate::assert_true(assign0 %in% c("score", "both", "none"), add = arg_checks)
   # score must be charlson, elixhauser; case insensitive
   score <- tolower(score)
   checkmate::assert_choice(score, choices = c("charlson", "elixhauser"), add = arg_checks)
   # icd must be icd9, icd10; case insensitive
   icd <- tolower(icd)
   checkmate::assert_choice(icd, choices = c("icd9", "icd10"), add = arg_checks)
-  # assign0, factorise, labelled, tidy.codes, parallel must be a single boolean value
-  checkmate::assert_logical(assign0, len = 1, add = arg_checks)
+  # factorise, labelled, tidy.codes, parallel must be a single boolean value
   checkmate::assert_logical(factorise, len = 1, add = arg_checks)
   checkmate::assert_logical(labelled, len = 1, add = arg_checks)
   checkmate::assert_logical(tidy.codes, len = 1, add = arg_checks)
@@ -195,17 +219,36 @@ comorbidity <- function(x, id, code, score, assign0, icd = "icd10", factorise = 
 
   ### Compute Charlson score and Charlson index
   if (score == "charlson") {
-    x$score <- with(x, ami + chf + pvd + cevd + dementia + copd + rheumd + pud + mld * ifelse(msld == 1 & assign0, 0, 1) + diab * ifelse(diabwc == 1 & assign0, 0, 1) + diabwc + hp + rend + canc * ifelse(metacanc == 1 & assign0, 0, 1) + msld + metacanc + aids)
+    x$score <- with(x, ami + chf + pvd + cevd + dementia + copd + rheumd + pud + mld * ifelse(msld == 1 & assign0 != "none", 0, 1) + diab * ifelse(diabwc == 1 & assign0 != "none", 0, 1) + diabwc + hp + rend + canc * ifelse(metacanc == 1 & assign0 != "none", 0, 1) + msld + metacanc + aids)
     x$index <- with(x, cut(score, breaks = c(0, 1, 2.5, 4.5, Inf), labels = c("0", "1-2", "3-4", ">=5"), right = FALSE))
-    x$wscore <- with(x, ami + chf + pvd + cevd + dementia + copd + rheumd + pud + mld * ifelse(msld == 1 & assign0, 0, 1) + diab * ifelse(diabwc == 1 & assign0, 0, 1) + diabwc * 2 + hp * 2 + rend * 2 + canc * ifelse(metacanc == 1 & assign0, 0, 2) + msld * 3 + metacanc * 6 + aids * 6)
+    x$wscore <- with(x, ami + chf + pvd + cevd + dementia + copd + rheumd + pud + mld * ifelse(msld == 1 & assign0 != "none", 0, 1) + diab * ifelse(diabwc == 1 & assign0 != "none", 0, 1) + diabwc * 2 + hp * 2 + rend * 2 + canc * ifelse(metacanc == 1 & assign0 != "none", 0, 2) + msld * 3 + metacanc * 6 + aids * 6)
     x$windex <- with(x, cut(wscore, breaks = c(0, 1, 2.5, 4.5, Inf), labels = c("0", "1-2", "3-4", ">=5"), right = FALSE))
   } else {
-    x$score <- with(x, chf + carit + valv + pcd + pvd + hypunc * ifelse(hypc == 1 & assign0, 0, 1) + hypc + para + ond + cpd + diabunc * ifelse(diabc == 1 & assign0, 0, 1) + diabc + hypothy + rf + ld + pud + aids + lymph + metacanc + solidtum * ifelse(metacanc == 1 & assign0, 0, 1) + rheumd + coag + obes + wloss + fed + blane + dane + alcohol + drug + psycho + depre)
+    x$score <- with(x, chf + carit + valv + pcd + pvd + hypunc * ifelse(hypc == 1 & assign0 != "none", 0, 1) + hypc + para + ond + cpd + diabunc * ifelse(diabc == 1 & assign0 != "none", 0, 1) + diabc + hypothy + rf + ld + pud + aids + lymph + metacanc + solidtum * ifelse(metacanc == 1 & assign0 != "none", 0, 1) + rheumd + coag + obes + wloss + fed + blane + dane + alcohol + drug + psycho + depre)
     x$index <- with(x, cut(score, breaks = c(-Inf, 0, 1, 4.5, Inf), labels = c("<0", "0", "1-4", ">=5"), right = FALSE))
-    x$wscore_ahrq <- with(x, chf * 9 + carit * 0 + valv * 0 + pcd * 6 + pvd * 3 + ifelse(hypunc == 1 | hypc == 1, 1, 0) * (-1) + para * 5 + ond * 5 + cpd * 3 + diabunc * ifelse(diabc == 1 & assign0, 0, 0) + diabc * (-3) + hypothy * 0 + rf * 6 + ld * 4 + pud * 0 + aids * 0 + lymph * 6 + metacanc * 14 + solidtum * ifelse(metacanc == 1 & assign0, 0, 7) + rheumd * 0 + coag * 11 + obes * (-5) + wloss * 9 + fed * 11 + blane * (-3) + dane * (-2) + alcohol * (-1) + drug * (-7) + psycho * (-5) + depre * (-5))
-    x$wscore_vw <- with(x, chf * 7 + carit * 5 + valv * (-1) + pcd * 4 + pvd * 2 + ifelse(hypunc == 1 | hypc == 1, 1, 0) * 0 + para * 7 + ond * 6 + cpd * 3 + diabunc * ifelse(diabc == 1 & assign0, 0, 0) + diabc * 0 + hypothy * 0 + rf * 5 + ld * 11 + pud * 0 + aids * 0 + lymph * 9 + metacanc * 12 + solidtum * ifelse(metacanc == 1 & assign0, 0, 4) + rheumd * 0 + coag * 3 + obes * (-4) + wloss * 6 + fed * 5 + blane * (-2) + dane * (-2) + alcohol * 0 + drug * (-7) + psycho * 0 + depre * (-3))
+    x$wscore_ahrq <- with(x, chf * 9 + carit * 0 + valv * 0 + pcd * 6 + pvd * 3 + ifelse(hypunc == 1 | hypc == 1, 1, 0) * (-1) + para * 5 + ond * 5 + cpd * 3 + diabunc * ifelse(diabc == 1 & assign0 != "none", 0, 0) + diabc * (-3) + hypothy * 0 + rf * 6 + ld * 4 + pud * 0 + aids * 0 + lymph * 6 + metacanc * 14 + solidtum * ifelse(metacanc == 1 & assign0 != "none", 0, 7) + rheumd * 0 + coag * 11 + obes * (-5) + wloss * 9 + fed * 11 + blane * (-3) + dane * (-2) + alcohol * (-1) + drug * (-7) + psycho * (-5) + depre * (-5))
+    x$wscore_vw <- with(x, chf * 7 + carit * 5 + valv * (-1) + pcd * 4 + pvd * 2 + ifelse(hypunc == 1 | hypc == 1, 1, 0) * 0 + para * 7 + ond * 6 + cpd * 3 + diabunc * ifelse(diabc == 1 & assign0 != "none", 0, 0) + diabc * 0 + hypothy * 0 + rf * 5 + ld * 11 + pud * 0 + aids * 0 + lymph * 9 + metacanc * 12 + solidtum * ifelse(metacanc == 1 & assign0 != "none", 0, 4) + rheumd * 0 + coag * 3 + obes * (-4) + wloss * 6 + fed * 5 + blane * (-2) + dane * (-2) + alcohol * 0 + drug * (-7) + psycho * 0 + depre * (-3))
     x$windex_ahrq <- with(x, cut(wscore_ahrq, breaks = c(-Inf, 0, 1, 4.5, Inf), labels = c("<0", "0", "1-4", ">=5"), right = FALSE))
     x$windex_vw <- with(x, cut(wscore_vw, breaks = c(-Inf, 0, 1, 4.5, Inf), labels = c("<0", "0", "1-4", ">=5"), right = FALSE))
+  }
+
+  ### If 'assign0 = "both"', then apply hierarchy to individual comorbidity domains too
+  if (assign0 == "both") {
+    if (score == "charlson") {
+      # "Mild liver disease" (`mld`) and "Moderate/severe liver disease" (`msld`)
+      x$mld[x$msld == 1] <- 0
+      # "Diabetes" (`diab`) and "Diabetes with complications" (`diabwc`)
+      x$diab[x$diabwc == 1] <- 0
+      # "Cancer" (`canc`) and "Metastatic solid tumour" (`metacanc`)
+      x$canc[x$metacanc == 1] <- 0
+    } else {
+      # "Hypertension, uncomplicated" (`hypunc`) and "Hypertension, complicated" (`hypc`)
+      x$hypunc[x$hypc == 1] <- 0
+      # "Diabetes, uncomplicated" (`diabunc`) and "Diabetes, complicated" (`diabc`)
+      x$diabunc[x$diabc == 1] <- 0
+      # "Solid tumour" (`solidtum`) and "Metastatic cancer" (`metacanc`)
+      x$solidtum[x$metacanc == 1] <- 0
+    }
   }
 
   ### Check output for possible unknown-state errors