+++ tests/testthat/test-comorbidity.R
@@ -84,7 +84,7 @@ test_that("charlson checks for its arguments properly", {
     comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = 1)
   )
   expect_error(
-    comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "1")
+    comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = TRUE)
   )
   expect_error(
     comorbidity(x = x, id = "id", code = "code", score = "charlson", factorise = 1)
@@ -138,13 +138,13 @@ test_that("comorbidity returns a data.frame", {
     code = sample_diag(50, version = "ICD9_2015"),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
   expect_s3_class(cs, "data.frame")
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = "none")
   expect_s3_class(cs, "data.frame")
-  cs <- comorbidity(x = x9, id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  cs <- comorbidity(x = x9, id = "id", code = "code", score = "charlson", assign0 = "none")
   expect_s3_class(cs, "data.frame")
-  cs <- comorbidity(x = x9, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
+  cs <- comorbidity(x = x9, id = "id", code = "code", score = "elixhauser", assign0 = "none")
   expect_s3_class(cs, "data.frame")
 })
 
@@ -154,14 +154,14 @@ test_that("comorbidity returns a data.frame with the correct number of rows", {
     code = sample_diag(10 * 5),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
   expect_equal(nrow(cs), 5)
   x <- data.frame(
     id = sample(1:50, size = 10 * 50, replace = TRUE),
     code = sample_diag(10 * 50, version = "ICD9_2015"),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
   expect_equal(nrow(cs), 50)
 })
 
@@ -171,9 +171,9 @@ test_that("if labelled = TRUE, comorbidity returns variable labels", {
     code = sample_diag(10 * 5),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", labelled = TRUE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", labelled = TRUE, assign0 = "none")
   expect_false(is.null(attr(cs, "variable.labels")))
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", labelled = TRUE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", labelled = TRUE, assign0 = "none")
   expect_false(is.null(attr(cs, "variable.labels")))
 })
 
@@ -183,22 +183,22 @@ test_that("if labelled = FALSE, comorbidity does not return variable labels", {
     code = sample_diag(10 * 5),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", labelled = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", labelled = FALSE, assign0 = "none")
   expect_true(is.null(attr(cs, "variable.labels")))
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", labelled = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", labelled = FALSE, assign0 = "none")
   expect_true(is.null(attr(cs, "variable.labels")))
   x9 <- data.frame(
     id = sample(1:5, size = 10 * 5, replace = TRUE),
     code = sample_diag(10 * 5, version = "ICD9_2015"),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", labelled = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", labelled = FALSE, assign0 = "none")
   expect_true(is.null(attr(cs, "variable.labels")))
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", labelled = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", labelled = FALSE, assign0 = "none")
   expect_true(is.null(attr(cs, "variable.labels")))
-  cs <- comorbidity(x = x9, id = "id", code = "code", score = "charlson", labelled = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x9, id = "id", code = "code", score = "charlson", labelled = FALSE, assign0 = "none")
   expect_true(is.null(attr(cs, "variable.labels")))
-  cs <- comorbidity(x = x9, id = "id", code = "code", score = "elixhauser", labelled = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x9, id = "id", code = "code", score = "elixhauser", labelled = FALSE, assign0 = "none")
   expect_true(is.null(attr(cs, "variable.labels")))
 })
 
@@ -208,7 +208,7 @@ test_that("if factorise = TRUE comorbidity returns factors", {
     code = sample_diag(10 * 5),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", factorise = TRUE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", factorise = TRUE, assign0 = "none")
   expect_s3_class(cs$ami, "factor")
   expect_s3_class(cs$chf, "factor")
   expect_s3_class(cs$pvd, "factor")
@@ -228,7 +228,7 @@ test_that("if factorise = TRUE comorbidity returns factors", {
   expect_s3_class(cs$aids, "factor")
   expect_s3_class(cs$index, "factor")
   expect_s3_class(cs$windex, "factor")
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = TRUE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = TRUE, assign0 = "none")
   expect_s3_class(cs$chf, "factor")
   expect_s3_class(cs$carit, "factor")
   expect_s3_class(cs$valv, "factor")
@@ -268,7 +268,7 @@ test_that("if factorise = TRUE comorbidity returns factors", {
     code = sample_diag(10 * 5, version = "ICD9_2015"),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", factorise = TRUE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "charlson", factorise = TRUE, assign0 = "none")
   expect_s3_class(cs$ami, "factor")
   expect_s3_class(cs$chf, "factor")
   expect_s3_class(cs$pvd, "factor")
@@ -288,7 +288,7 @@ test_that("if factorise = TRUE comorbidity returns factors", {
   expect_s3_class(cs$aids, "factor")
   expect_s3_class(cs$index, "factor")
   expect_s3_class(cs$windex, "factor")
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = TRUE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = TRUE, assign0 = "none")
   expect_s3_class(cs$chf, "factor")
   expect_s3_class(cs$carit, "factor")
   expect_s3_class(cs$valv, "factor")
@@ -331,7 +331,7 @@ test_that("if factorise = FALSE comorbidity does not return factors", {
     code = sample_diag(10 * 5),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", score = "charlson", code = "code", factorise = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", score = "charlson", code = "code", factorise = FALSE, assign0 = "none")
   expect_false("factor" %in% class(cs$ami))
   expect_false("factor" %in% class(cs$chf))
   expect_false("factor" %in% class(cs$pvd))
@@ -351,7 +351,7 @@ test_that("if factorise = FALSE comorbidity does not return factors", {
   expect_false("factor" %in% class(cs$aids))
   expect_s3_class(cs$index, "factor")
   expect_s3_class(cs$windex, "factor")
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = FALSE, assign0 = "none")
   expect_false("factor" %in% class(cs$chf))
   expect_false("factor" %in% class(cs$carit))
   expect_false("factor" %in% class(cs$valv))
@@ -391,7 +391,7 @@ test_that("if factorise = FALSE comorbidity does not return factors", {
     code = sample_diag(10 * 5, version = "ICD9_2015"),
     stringsAsFactors = FALSE
   )
-  cs <- comorbidity(x = x, id = "id", score = "charlson", code = "code", factorise = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", score = "charlson", code = "code", factorise = FALSE, assign0 = "none")
   expect_false("factor" %in% class(cs$ami))
   expect_false("factor" %in% class(cs$chf))
   expect_false("factor" %in% class(cs$pvd))
@@ -411,7 +411,7 @@ test_that("if factorise = FALSE comorbidity does not return factors", {
   expect_false("factor" %in% class(cs$aids))
   expect_s3_class(cs$index, "factor")
   expect_s3_class(cs$windex, "factor")
-  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = FALSE, assign0 = FALSE)
+  cs <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", factorise = FALSE, assign0 = "none")
   expect_false("factor" %in% class(cs$chf))
   expect_false("factor" %in% class(cs$carit))
   expect_false("factor" %in% class(cs$valv))
@@ -453,7 +453,7 @@ test_that("comorbidity scores are 0 or 1", {
     record_id = 8,
     diagnosis_icd_2 = c("G92", "J984, M419", "A0472", "A419", "D696", "E11621", "E119", "E669", "E875", "G40909", "G8220", "G904", "I10", "I248", "I2720", "I5030", "I5033", "I82411", "I824Y1", "J45909", "J9621", "K5900", "K592", "L89310", "L89620", "L89890", "L97529", "M21372", "M419", "N179", "N319", "N390", "Q056", "Q058", "R440", "R578", "T426X5A", "T428X5A", "T83510A", "Z6835", "Z713", "Z781", "Z7982", "Z853", "Z981")
   )
-  elixhauser10 <- comorbidity(x = dat, id = "record_id", code = "diagnosis_icd_2", score = "elixhauser", icd = "icd10", assign0 = T, factorise = F, labelled = F, tidy.codes = F)[, 2:32]
+  elixhauser10 <- comorbidity(x = dat, id = "record_id", code = "diagnosis_icd_2", score = "elixhauser", icd = "icd10", assign0 = "score", factorise = F, labelled = F, tidy.codes = F)[, 2:32]
   expect_true(object = all(elixhauser10 >= 0 & elixhauser10 <= 1))
 
   for (i in seq(25)) {
@@ -462,18 +462,18 @@ test_that("comorbidity scores are 0 or 1", {
       code = sample_diag(50),
       stringsAsFactors = FALSE
     )
-    elixhauser10 <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", icd = "icd10", assign0 = FALSE)[, 2:32]
+    elixhauser10 <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", icd = "icd10", assign0 = "none")[, 2:32]
     expect_true(object = all(elixhauser10 >= 0 & elixhauser10 <= 1))
-    charlson10 <- comorbidity(x = x, id = "id", code = "code", score = "charlson", icd = "icd10", assign0 = FALSE)[, 2:18]
+    charlson10 <- comorbidity(x = x, id = "id", code = "code", score = "charlson", icd = "icd10", assign0 = "none")[, 2:18]
     expect_true(object = all(charlson10 >= 0 & charlson10 <= 1))
     x <- data.frame(
       id = sample(1:5, size = 50, replace = TRUE),
       code = sample_diag(50, version = "ICD9_2015"),
       stringsAsFactors = FALSE
     )
-    elixhauser9 <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", icd = "icd9", assign0 = FALSE)[, 2:32]
+    elixhauser9 <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", icd = "icd9", assign0 = "none")[, 2:32]
     expect_true(object = all(elixhauser9 >= 0 & elixhauser9 <= 1))
-    charlson9 <- comorbidity(x = x, id = "id", code = "code", score = "charlson", icd = "icd9", assign0 = FALSE)[, 2:18]
+    charlson9 <- comorbidity(x = x, id = "id", code = "code", score = "charlson", icd = "icd9", assign0 = "none")[, 2:18]
     expect_true(object = all(charlson9 >= 0 & charlson9 <= 1))
   }
 })
@@ -488,14 +488,14 @@ test_that("duplicate codes are not counted twice (or more)", {
     x2 <- rbind(x, x)
     x3 <- rbind(x, x, x)
     x4 <- rbind(x, x, x, x)
-    cx <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
-    cx2 <- comorbidity(x = x2, id = "id", code = "code", score = "charlson", assign0 = FALSE)
-    cx3 <- comorbidity(x = x3, id = "id", code = "code", score = "charlson", assign0 = FALSE)
-    cx4 <- comorbidity(x = x4, id = "id", code = "code", score = "charlson", assign0 = FALSE)
-    ex <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
-    ex2 <- comorbidity(x = x2, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
-    ex3 <- comorbidity(x = x3, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
-    ex4 <- comorbidity(x = x4, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
+    cx <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
+    cx2 <- comorbidity(x = x2, id = "id", code = "code", score = "charlson", assign0 = "none")
+    cx3 <- comorbidity(x = x3, id = "id", code = "code", score = "charlson", assign0 = "none")
+    cx4 <- comorbidity(x = x4, id = "id", code = "code", score = "charlson", assign0 = "none")
+    ex <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = "none")
+    ex2 <- comorbidity(x = x2, id = "id", code = "code", score = "elixhauser", assign0 = "none")
+    ex3 <- comorbidity(x = x3, id = "id", code = "code", score = "elixhauser", assign0 = "none")
+    ex4 <- comorbidity(x = x4, id = "id", code = "code", score = "elixhauser", assign0 = "none")
     expect_equal(object = cx, expected = cx2)
     expect_equal(object = cx, expected = cx3)
     expect_equal(object = cx, expected = cx4)
@@ -528,14 +528,14 @@ test_that("duplicate codes are not counted twice (or more)", {
     x2 <- rbind(x, x)
     x3 <- rbind(x, x, x)
     x4 <- rbind(x, x, x, x)
-    cx <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = FALSE)
-    cx2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = FALSE)
-    cx3 <- comorbidity(x = x3, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = FALSE)
-    cx4 <- comorbidity(x = x4, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = FALSE)
-    ex <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = FALSE)
-    ex2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = FALSE)
-    ex3 <- comorbidity(x = x3, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = FALSE)
-    ex4 <- comorbidity(x = x4, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = FALSE)
+    cx <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = "none")
+    cx2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = "none")
+    cx3 <- comorbidity(x = x3, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = "none")
+    cx4 <- comorbidity(x = x4, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = "none")
+    ex <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = "none")
+    ex2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = "none")
+    ex3 <- comorbidity(x = x3, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = "none")
+    ex4 <- comorbidity(x = x4, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = "none")
     expect_equal(object = cx, expected = cx2)
     expect_equal(object = cx, expected = cx3)
     expect_equal(object = cx, expected = cx4)
@@ -567,11 +567,11 @@ test_that("input dataset with additional columns", {
   )
   x2 <- x
   x2$noise <- rnorm(n = nrow(x2))
-  c <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
-  c2 <- comorbidity(x = x2, id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  c <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
+  c2 <- comorbidity(x = x2, id = "id", code = "code", score = "charlson", assign0 = "none")
   expect_equal(object = c2, expected = c)
-  e <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
-  e2 <- comorbidity(x = x2, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
+  e <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = "none")
+  e2 <- comorbidity(x = x2, id = "id", code = "code", score = "elixhauser", assign0 = "none")
   expect_equal(object = e2, expected = e)
 
   x <- data.frame(
@@ -582,31 +582,31 @@ test_that("input dataset with additional columns", {
   )
   x2 <- x
   x2$noise <- rnorm(n = nrow(x2))
-  c <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = FALSE)
-  c2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = FALSE)
+  c <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = "none")
+  c2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "charlson", assign0 = "none")
   expect_equal(object = c2, expected = c)
-  e <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = FALSE)
-  e2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = FALSE)
+  e <- comorbidity(x = x, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = "none")
+  e2 <- comorbidity(x = x2, id = "id", code = "code", icd = "icd9", score = "elixhauser", assign0 = "none")
   expect_equal(object = e2, expected = e)
 })
 
 test_that("all comorbidities", {
   data("icd10_2011", package = "comorbidity")
   icd10_2011$id <- 1
-  c <- comorbidity(x = icd10_2011, id = "id", code = "Code", icd = "icd10", score = "charlson", assign0 = T)
-  e <- comorbidity(x = icd10_2011, id = "id", code = "Code", icd = "icd10", score = "elixhauser", assign0 = T)
+  c <- comorbidity(x = icd10_2011, id = "id", code = "Code", icd = "icd10", score = "charlson", assign0 = "score")
+  e <- comorbidity(x = icd10_2011, id = "id", code = "Code", icd = "icd10", score = "elixhauser", assign0 = "score")
   expect_true(object = all(c[2:18] == 1))
   expect_true(object = all(e[2:32] == 1))
   data("icd10_2009", package = "comorbidity")
   icd10_2009$id <- 1
-  c <- comorbidity(x = icd10_2009, id = "id", code = "Code", icd = "icd10", score = "charlson", assign0 = T)
-  e <- comorbidity(x = icd10_2009, id = "id", code = "Code", icd = "icd10", score = "elixhauser", assign0 = T)
+  c <- comorbidity(x = icd10_2009, id = "id", code = "Code", icd = "icd10", score = "charlson", assign0 = "score")
+  e <- comorbidity(x = icd10_2009, id = "id", code = "Code", icd = "icd10", score = "elixhauser", assign0 = "score")
   expect_true(object = all(c[2:18] == 1))
   expect_true(object = all(e[2:32] == 1))
   data("icd9_2015", package = "comorbidity")
   icd9_2015$id <- 1
-  c <- comorbidity(x = icd9_2015, id = "id", code = "Code", icd = "icd9", score = "charlson", assign0 = T)
-  e <- comorbidity(x = icd9_2015, id = "id", code = "Code", icd = "icd9", score = "elixhauser", assign0 = T)
+  c <- comorbidity(x = icd9_2015, id = "id", code = "Code", icd = "icd9", score = "charlson", assign0 = "score")
+  e <- comorbidity(x = icd9_2015, id = "id", code = "Code", icd = "icd9", score = "elixhauser", assign0 = "score")
   expect_true(object = all(c[2:18] == 1))
   expect_true(object = all(e[2:32] == 1))
 })
@@ -617,10 +617,10 @@ test_that("break output checks", {
     code = sample_diag(50),
     stringsAsFactors = FALSE
   )
-  cx <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  cx <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
   cx[2:18] <- cx[2:18] + rnorm(n = nrow(cx))
   expect_error(.check_output(x = cx, id = "id", score = "charlson"), regexp = "unexpected state")
-  ex <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = FALSE)
+  ex <- comorbidity(x = x, id = "id", code = "code", score = "elixhauser", assign0 = "none")
   ex[2:32] <- ex[2:32] + rnorm(n = nrow(ex))
   expect_error(.check_output(x = ex, id = "id", score = "elixhauser"), regexp = "unexpected state")
 })
@@ -632,8 +632,8 @@ test_that("works ok with data.table", {
     code = sample_diag(50),
     stringsAsFactors = FALSE
   )
-  c <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
-  c.dt <- comorbidity(x = data.table::setDT(x), id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  c <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
+  c.dt <- comorbidity(x = data.table::setDT(x), id = "id", code = "code", score = "charlson", assign0 = "none")
   expect_equal(object = c.dt, expected = c)
   x <- data.frame(
     id = sample(1:20, size = 50, replace = TRUE),
@@ -641,7 +641,7 @@ test_that("works ok with data.table", {
     stringsAsFactors = FALSE
   )
   x$noise <- stats::rnorm(n = nrow(x))
-  c <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = FALSE)
-  c.dt <- comorbidity(x = data.table::setDT(x), id = "id", code = "code", score = "charlson", assign0 = FALSE)
+  c <- comorbidity(x = x, id = "id", code = "code", score = "charlson", assign0 = "none")
+  c.dt <- comorbidity(x = data.table::setDT(x), id = "id", code = "code", score = "charlson", assign0 = "none")
   expect_equal(object = c.dt, expected = c)
 })